Move code for reading process regions into input/process/region.rs
